requirements.txt
pyTelegramBotAPI==4.23.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1